scriptcomp: Remove m_nDebugSymbolicOutput and everything about it (#71)
It has bitrotten and is not quite accurate. Plus, it makes future
additions harder as you need to implement them in both branches.

Instead, we should always be generating binary NCS and then
disassembling it to see the text output.

This should all be deadcode. I _think_ I can also remove
`m_nBinaryCodeLength` as that should be the same as
`m_nOutputCodeLength` now, but I'll defer that to another PR.

Test suite passes, as do some other ad hoc test scripts I had. The
generated NCS is the same as it was.

---
This patch is released as a dual MIT and WTFPL-2 license :)