scriptcomp: Optimize typical for loop construct instructions (#86)
The common loop pattern of `for (i = 0; i < 1000; ++i)` gets compiled
into
```
     8  RUNSTACK_ADD, TYPE_STRING               
    10  CONSTANT, TYPE_INTEGER, 0               
    16  RUNSTACK_COPY, TYPE_VOID, -4, 4         
    24  CONSTANT, TYPE_INTEGER, 1000            
    30  LT, TYPETYPE_INTEGER_INTEGER            
    32  JZ, 32                                  
    38  INCREMENT, TYPE_INTEGER, -4             
    44  RUNSTACK_COPY, TYPE_VOID, -4, 4         
    52  MODIFY_STACK_POINTER, -4                
    58  JMP, -42                                
```
The inc/dec operations can reach all across the stack and don't need to
be copied over to top of stack. However, since `++i` returns a value we
have to emit the RUNSTACK_COPY. But, when we don't use that value, we
just immediately clear it with MODIFY_STACK_POINTER. Just don't do
either.

## Testing

Added an extra test case with ++i and runtime check on iteration count.
Verified disassembly of
```nss
void main() {
    int i;
    for (i = 0; i < 1000; ++i) { }
}
```
is sane.

## Changelog

### Performance Improvements
- Optimized the common for loop construct `for (i = 0; i < N; ++i)`

## Licence

- [x] I am licencing my change under the project's MIT licence,
including all changes to GPL-3.0 licenced parts of the codebase.