Fix checksums import when used as library
The previous fix to checksums that tried to restore Nim 1.6.x support
allowed the neverwinter tools to build. However, the `private` folder is
not accessible when using neverwinter.nim as a library, so compilation
of other tools now breaks. This fix fixes the fix so that it actually
fixes 1.6.x support.