Fix checksums import when used as library (#110)
The previous fix to checksums that tried to restore Nim 1.6.x support
allowed the neverwinter tools to build. However, the `private` folder is
not accessible when using neverwinter.nim as a library, so compilation
of other tools now breaks. This fix fixes the fix so that it *actually*
fixes 1.6.x support.

## Testing

Compiles on Nim 1.6.14 and 2.0.2, both when installing binaries and when
included as a library (e.g., when building nasher).

## Licence

- [x] I am licencing my change under the project's MIT licence,
including all changes to GPL-3.0 licenced parts of the codebase.