scriptcomp: hack in FUNCTION, FILE, DATE, TIME macros (#113)
Current implementation disallow use of FUNC, FILE and LINE as default
arguments to functions. Author would probably expect them to be filled
in at the callsite; the current behaviour would be incorrect. So this
has been deferred until the later compiler can be taught to backfill
this.

## Testing

Included testcase passes.

## Changelog

### Added

- scriptcomp: Added support for `__FUNCTION__`, `__FILE__`,`__LINE__`,
`__DATE__` and `__TIME__`.

## Licence

- [x] I am licencing my change under the project's MIT licence,
including all changes to GPL-3.0 licenced parts of the codebase.